+++ cmd/serverless-sidecar/.gitignore
@@ -1,1 +1,1 @@
-serverless-init
+serverless-sidecar
