+++ cmd/serverless-sidecar/cloudservice/cloudrun.go
@@ -8,7 +8,7 @@ package cloudservice
 import (
 	"os"
 
-	"github.com/DataDog/datadog-agent/cmd/serverless-init/cloudservice/helper"
+	"github.com/DataDog/datadog-agent/cmd/serverless-sidecar/cloudservice/helper"
 )
 
 const (
